requirements.txt
alembic==1.13.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distlib==0.3.8
filelock==3.16.1
Flask==3.0.3
Flask-Admin==1.6.1
Flask-JWT==0.3.2
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==2.1.5
nodeenv==1.9.1
packaging==24.1
pipdeptree==2.23.4
pipenv==2024.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyright==1.1.385
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
SQLAlchemy-serializer==1.4.12
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.6
Werkzeug==3.0.4
WTForms==3.1.2
zipp==3.20.2
